update rebuild.py to build dash TOC tags
we use html5lib with BeautifulSoup, since it seems to be the least bad
at munging the rest of the document.  html.parser and lxml both break
things to the point that the header at the top of each RFC does not
display correctly.  html5lib still munges some things (like reformating
markup and reordering attributes), but they do not appear to affect the
display of the page.

also update Info.plist to look for tables of content.