Rewrite rebuild script
It now makes use of the JSON metadata files to get the title of each document,
instead of having to parse the HTML. This means less parsing is needed, which
should make the project easier to maintain going forwards.

This also removes the `index.html` file. It might make sense to generate this
when going through all the RFCs, but to make things easier for now, I've
decided to scrap it.