Bring content up to date (#11)
This updates the scripts in the repository so they can be used to pull in new RFCs, fixing https://github.com/willnorris/rfcdash/issues/10.

I've changed the `sync` script to retrieve RFCs via rsync, which seems like a much more efficient way than scraping the website.

The `rebuild.py` script has then been updated to match the new structure of where files are located, and to read the RFC titles from the JSON metadata files instead of trying to parse it out from the HTML files.

I haven't made any adjustments to how the new HTML format for RFCs is handled, as they seem to work well as-is when loaded into Dash.

As part of this I've removed the `index.html` file. I'm not sure it really helped with anything before, and it's not clear to me what a current alternative would look like.